feat(api): cancel, rearrange and missed site visit logic for linking (A2-7499) (A2-7480) (A2-7474)
Split out the tests to multiple files and updated to cover linked multiple appellants
