+++ appeals/api/src/server/repositories/site-visit.repository.js
@@ -39,11 +39,40 @@ const updateSiteVisitById = (id, data) =>
 	});
 
 /**
+ * @param {number[]} appealIds
+ * @param {{
+ * 	visitDate?: Date;
+ * 	visitEndTime?: Date;
+ * 	visitStartTime?: Date;
+ * 	siteVisitTypeId?: number;
+ * 	whoMissedSiteVisit?: string
+ * }} data
+ * @returns {PrismaPromise<object>}
+ */
+const updateMultiSiteVisitByAppealId = (appealIds, data) =>
+	databaseConnector.siteVisit.updateMany({
+		where: {
+			appealId: { in: appealIds }
+		},
+		data
+	});
+
+/**
  * @param {number} id
  * */
 const deleteSiteVisitById = (id) =>
 	databaseConnector.siteVisit.delete({
 		where: { id }
+	});
+
+/**
+ * @param {number[]} appealIds
+ * */
+const deleteMultiSiteVisitByAppealId = (appealIds) =>
+	databaseConnector.siteVisit.deleteMany({
+		where: {
+			appealId: { in: appealIds }
+		}
 	});
 
 /**
@@ -84,7 +113,9 @@ const getSiteVisitById = (id) => {
 export default {
 	createSiteVisitById,
 	updateSiteVisitById,
+	updateMultiSiteVisitByAppealId,
 	deleteSiteVisitById,
+	deleteMultiSiteVisitByAppealId,
 	getMissedSiteVisitByAppealId,
 	getSiteVisitById
 };
